requirements.txt
python-telegram-bot
asyncpg
psycopg2
sqlalchemy
pydantic
alembic
alembic_utils
python-dateutil